Fix env vars from JenkinsPipelineBuildStrategy could be null values
When start a JenkinsPipelineBuildStrategy with `oc start-build`, if
there is an environment variable with an empty string value, the
corresponding parameter will be set to null value. This could lead to
exceptions.

This PR fixes this issue by replacing null values with empty strings
when passing environment variables from JenkinsPipelineBuildStrategy to
ParametersAction.